Feat: [BE] 댓글 수정/ 삭제 기능 #131 #132

+++ BE/urida/urida-board/src/main/java/com/urida/comment/service/CommentService.java
@@ -15,6 +15,6 @@ public interface CommentService {
 
     Comment updateComment(Long comment_id, CommentRequestDto commentRequestDto);
 
-    int removeComment(Long comment_id);
+    Long removeComment(Long comment_id);
 
 }
